+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simple and Rolling Cumulative Sum
 - Stop-Loss Logic: High-to-Low, Buy-to-Low
 - Add Factor Datasource
+- Calculate Factor betas for each allocation in strategies
 - Add MSCI Issuerid to final df 
 ### Fixed
 ### Changed